requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
streamlit>=1.32.0
nltk
spacy
tqdm
plotly
PyPDF2
altair>=5.0.0,<7.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
openai>=1.40.0
python-dotenv>=1.0.0
requests>=2.31.0
